refactor(html_extractor): delegate text, heading, and ID extraction to dedicated classes under HtmlExtractor namespace
